findmy_mqtt/requirements.txt
findmy==0.9.8
paho-mqtt>=1.6,<3